requirements.txt
numpy>=1.19.5
pandas>=1.1.5
requests>=2.27.1
scikit-learn~=0.22.2.post1
pymongo~=3.12.0
pyyaml~=5.4.1
flask~=2.0.3
dataclasses~=0.6
prometheus_client~=0.13.1
werkzeug~=2.0.3